feat(build): #16 enhance try cognitive complexity
- Upgrade the cognitive complexity when multiple
functions are defined in a nested way.
- Also, remove to add the nested level to all the
cases.
- Consider the case when multiple exceptions are
caught in a single try block.
- Update the test case to cover the new feature.
