+++ src/cognitive_complexity/mod.rs
@@ -153,12 +153,28 @@ fn statement_cognitive_complexity(statement: Stmt, nesting_level: u64) -> PyResu
     match statement {
         Stmt::FunctionDef(f) => {
             for node in f.body.iter() {
-                complexity += statement_cognitive_complexity(node.clone(), nesting_level)?;
+                match node {
+                    Stmt::FunctionDef(..) | Stmt::AsyncFunctionDef(..) => {
+                        complexity +=
+                            statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
+                    }
+                    _ => {
+                        complexity += statement_cognitive_complexity(node.clone(), nesting_level)?;
+                    }
+                }
             }
         }
         Stmt::AsyncFunctionDef(f) => {
             for node in f.body.iter() {
-                complexity += statement_cognitive_complexity(node.clone(), nesting_level)?;
+                match node {
+                    Stmt::FunctionDef(..) | Stmt::AsyncFunctionDef(..) => {
+                        complexity +=
+                            statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
+                    }
+                    _ => {
+                        complexity += statement_cognitive_complexity(node.clone(), nesting_level)?;
+                    }
+                }
             }
         }
         Stmt::ClassDef(c) => {
@@ -168,23 +184,26 @@ fn statement_cognitive_complexity(statement: Stmt, nesting_level: u64) -> PyResu
         }
         Stmt::Assign(a) => {
             complexity += count_bool_ops(*a.value);
+
+            if complexity > 0 {
+                complexity += nesting_level;
+            }
         }
-        // breaks in the linear flow
         Stmt::For(f) => {
-            complexity += 1;
+            complexity += 1 + nesting_level;
             for node in f.body.iter() {
                 complexity += statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
             }
         }
         Stmt::While(w) => {
-            complexity += 1;
+            complexity += 1 + nesting_level;
             complexity += count_bool_ops(*w.test);
             for node in w.body.iter() {
                 complexity += statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
             }
         }
         Stmt::If(i) => {
-            complexity += 1;
+            complexity += 1 + nesting_level;
             complexity += count_bool_ops(*i.test);
             for node in i.body.iter() {
                 complexity += statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
@@ -203,12 +222,12 @@ fn statement_cognitive_complexity(statement: Stmt, nesting_level: u64) -> PyResu
             }
         }
         Stmt::Try(t) => {
-            complexity += 1;
             for node in t.body.iter() {
                 complexity += statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
             }
 
             for handler in t.handlers.iter() {
+                complexity += 1;
                 match handler {
                     ast::ExceptHandler::ExceptHandler(e) => {
                         for node in e.body.iter() {
@@ -218,6 +237,18 @@ fn statement_cognitive_complexity(statement: Stmt, nesting_level: u64) -> PyResu
                     }
                 }
             }
+
+            for node in t.orelse.iter() {
+                complexity += statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
+            }
+
+            for node in t.finalbody.iter() {
+                complexity += statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
+            }
+
+            if complexity > 0 {
+                complexity += nesting_level;
+            }
         }
         Stmt::Match(m) => {
             for case in m.cases.iter() {
@@ -225,13 +256,13 @@ fn statement_cognitive_complexity(statement: Stmt, nesting_level: u64) -> PyResu
                     complexity += statement_cognitive_complexity(node.clone(), nesting_level + 1)?;
                 }
             }
+
+            if complexity > 0 {
+                complexity += nesting_level;
+            }
         }
         _ => {}
     };
-
-    if complexity > 0 {
-        complexity += nesting_level;
-    }
 
     Ok(complexity)
 }
